refactor(markdown): split catalog/AntCat parsers
