+++ app/services/markdowns/mentioned_users.rb
@@ -13,7 +13,7 @@ module Markdowns
     private
 
       def user_ids
-        content.scan(Markdowns::ParseAntcatHooks::USER_TAG_REGEX).flatten.uniq
+        content.scan(Markdowns::ParseAntcatTags::USER_TAG_REGEX).flatten.uniq
       end
   end
 end
