+++ app/services/markdowns/parse_all_tags.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+module Markdowns
+  class ParseAllTags
+    include ActionView::Helpers::SanitizeHelper
+    include Service
+
+    def initialize content, sanitize_content: true
+      @content = sanitize_content ? sanitize(content).to_str : content
+    end
+
+    def call
+      parsed = Markdowns::ParseAntcatTags[content, sanitize_content: false]
+      parsed = Markdowns::ParseCatalogTags[parsed, sanitize_content: false]
+
+      parsed
+    end
+
+    private
+
+      attr_reader :content
+  end
+end
